AI/app/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
Bottleneck==1.3.5
brotlipy==0.7.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.4
click==8.0.4
colorama==0.4.6
cryptography==41.0.3
efficientnet-pytorch==0.7.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fastapi==0.104.0
filelock==3.12.4
h11==0.12.0
idna==3.4
Jinja2==2.11.3
joblib==1.2.0
MarkupSafe==1.1.1
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.2
numexpr==2.8.7
numpy==1.22.4
openpyxl==3.0.10
pandas==1.4.4
Pillow==10.0.1
pip==23.3
pycparser==2.21
pydantic==2.4.1
pydantic_core==2.10.1
pyOpenSSL==23.2.0
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
requests==2.31.0
scipy==1.11.3
setuptools==68.0.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
threadpoolctl==2.2.0
typing_extensions==4.8.0
urllib3==1.26.16
uvicorn==0.20.0
wheel==0.41.2
pika==1.3.1

